Refactor verify-action-build.py into modular package with tests
Split the 3068-line monolithic script into 15 focused modules organized
by responsibility (GitHub client, action parsing, Docker building, diff
engines, security analysis, PR extraction, etc.), extracted the inline
Dockerfile to a standalone file, and added 112 unit tests.

Generated-by: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
