+++ utils/tests/verify_action_build/test_diff_node_modules.py
@@ -1,0 +1,122 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+from verify_action_build.diff_node_modules import diff_node_modules
+
+
+class TestDiffNodeModules:
+    def test_identical_modules(self, tmp_path):
+        orig = tmp_path / "original"
+        rebuilt = tmp_path / "rebuilt"
+        orig.mkdir()
+        rebuilt.mkdir()
+
+        (orig / "lodash").mkdir()
+        (rebuilt / "lodash").mkdir()
+        (orig / "lodash" / "index.js").write_text("module.exports = {}")
+        (rebuilt / "lodash" / "index.js").write_text("module.exports = {}")
+
+        result = diff_node_modules(
+            orig, rebuilt, "test", "repo", "a" * 40,
+        )
+        assert result is True
+
+    def test_different_file_content(self, tmp_path):
+        orig = tmp_path / "original"
+        rebuilt = tmp_path / "rebuilt"
+        orig.mkdir()
+        rebuilt.mkdir()
+
+        (orig / "lodash").mkdir()
+        (rebuilt / "lodash").mkdir()
+        (orig / "lodash" / "index.js").write_text("original")
+        (rebuilt / "lodash" / "index.js").write_text("different")
+
+        result = diff_node_modules(
+            orig, rebuilt, "test", "repo", "a" * 40,
+        )
+        assert result is False
+
+    def test_extra_package_in_original(self, tmp_path):
+        orig = tmp_path / "original"
+        rebuilt = tmp_path / "rebuilt"
+        orig.mkdir()
+        rebuilt.mkdir()
+
+        (orig / "lodash").mkdir()
+        (orig / "lodash" / "index.js").write_text("code")
+        (orig / "malicious-pkg").mkdir()
+        (orig / "malicious-pkg" / "index.js").write_text("evil")
+        (rebuilt / "lodash").mkdir()
+        (rebuilt / "lodash" / "index.js").write_text("code")
+
+        result = diff_node_modules(
+            orig, rebuilt, "test", "repo", "a" * 40,
+        )
+        assert result is False
+
+    def test_empty_directories(self, tmp_path):
+        orig = tmp_path / "original"
+        rebuilt = tmp_path / "rebuilt"
+        orig.mkdir()
+        rebuilt.mkdir()
+
+        result = diff_node_modules(
+            orig, rebuilt, "test", "repo", "a" * 40,
+        )
+        assert result is True
+
+    def test_noisy_files_ignored(self, tmp_path):
+        orig = tmp_path / "original"
+        rebuilt = tmp_path / "rebuilt"
+        orig.mkdir()
+        rebuilt.mkdir()
+
+        (orig / "lodash").mkdir()
+        (rebuilt / "lodash").mkdir()
+        (orig / "lodash" / "index.js").write_text("code")
+        (rebuilt / "lodash" / "index.js").write_text("code")
+        # .package-lock.json differs but should be ignored
+        (orig / ".package-lock.json").write_text("{}")
+        (rebuilt / ".package-lock.json").write_text('{"different": true}')
+
+        result = diff_node_modules(
+            orig, rebuilt, "test", "repo", "a" * 40,
+        )
+        assert result is True
+
+    def test_package_json_install_fields_ignored(self, tmp_path):
+        import json
+        orig = tmp_path / "original"
+        rebuilt = tmp_path / "rebuilt"
+        orig.mkdir()
+        rebuilt.mkdir()
+
+        (orig / "lodash").mkdir()
+        (rebuilt / "lodash").mkdir()
+
+        orig_pkg = {"name": "lodash", "version": "4.0.0", "_resolved": "https://old"}
+        rebuilt_pkg = {"name": "lodash", "version": "4.0.0", "_resolved": "https://new"}
+
+        (orig / "lodash" / "package.json").write_text(json.dumps(orig_pkg))
+        (rebuilt / "lodash" / "package.json").write_text(json.dumps(rebuilt_pkg))
+
+        result = diff_node_modules(
+            orig, rebuilt, "test", "repo", "a" * 40,
+        )
+        assert result is True
