+++ utils/verify_action_build/diff_node_modules.py
@@ -1,0 +1,158 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+"""Vendored node_modules comparison."""
+
+import hashlib
+import json
+from pathlib import Path
+
+from .console import console, link
+from .diff_display import show_colored_diff
+
+
+def diff_node_modules(
+    original_dir: Path, rebuilt_dir: Path, org: str, repo: str, commit_hash: str,
+) -> bool:
+    """Compare original vs rebuilt node_modules. Return True if they match."""
+    blob_url = f"https://github.com/{org}/{repo}/blob/{commit_hash}/node_modules"
+
+    # Metadata files that legitimately differ between installs
+    noisy_files = {".package-lock.json", ".yarn-integrity"}
+    noisy_dirs = {".cache", ".package-lock.json"}
+
+    def collect_files(base: Path) -> dict[Path, str]:
+        """Collect all files under base with their SHA256 hashes."""
+        result = {}
+        for f in sorted(base.rglob("*")):
+            if not f.is_file():
+                continue
+            rel = f.relative_to(base)
+            if rel.name in noisy_files:
+                continue
+            if any(part in noisy_dirs for part in rel.parts):
+                continue
+            result[rel] = hashlib.sha256(f.read_bytes()).hexdigest()
+        return result
+
+    console.print()
+    console.rule("[bold]Comparing vendored node_modules[/bold]")
+
+    with console.status("[dim]Hashing files...[/dim]"):
+        original_files = collect_files(original_dir)
+        rebuilt_files = collect_files(rebuilt_dir)
+
+    original_set = set(original_files.keys())
+    rebuilt_set = set(rebuilt_files.keys())
+
+    only_in_original = sorted(original_set - rebuilt_set)
+    only_in_rebuilt = sorted(rebuilt_set - original_set)
+    common = sorted(original_set & rebuilt_set)
+
+    original_packages = sorted({p.parts[0] for p in original_set if len(p.parts) > 1})
+    rebuilt_packages = sorted({p.parts[0] for p in rebuilt_set if len(p.parts) > 1})
+    pkg_only_orig = set(original_packages) - set(rebuilt_packages)
+    pkg_only_rebuilt = set(rebuilt_packages) - set(original_packages)
+
+    console.print(
+        f"  [dim]Original: {len(original_files)} files in {len(original_packages)} packages[/dim]"
+    )
+    console.print(
+        f"  [dim]Rebuilt:  {len(rebuilt_files)} files in {len(rebuilt_packages)} packages[/dim]"
+    )
+
+    all_match = True
+
+    if pkg_only_orig:
+        all_match = False
+        console.print(f"\n  [red]Packages only in original ({len(pkg_only_orig)}):[/red]")
+        for pkg in sorted(pkg_only_orig):
+            console.print(f"    [red]-[/red] {pkg}")
+
+    if pkg_only_rebuilt:
+        all_match = False
+        console.print(f"\n  [red]Packages only in rebuilt ({len(pkg_only_rebuilt)}):[/red]")
+        for pkg in sorted(pkg_only_rebuilt):
+            console.print(f"    [green]+[/green] {pkg}")
+
+    extra_in_orig = [f for f in only_in_original if f.parts[0] not in pkg_only_orig]
+    if extra_in_orig:
+        all_match = False
+        console.print(f"\n  [red]Files only in original (not from extra packages) — {len(extra_in_orig)}:[/red]")
+        for f in extra_in_orig[:20]:
+            file_link = link(f"{blob_url}/{f}", str(f))
+            console.print(f"    [red]-[/red] {file_link}")
+        if len(extra_in_orig) > 20:
+            console.print(f"    [dim]... and {len(extra_in_orig) - 20} more[/dim]")
+
+    extra_in_rebuilt = [f for f in only_in_rebuilt if f.parts[0] not in pkg_only_rebuilt]
+    if extra_in_rebuilt:
+        console.print(f"\n  [yellow]Files only in rebuilt (not from extra packages) — {len(extra_in_rebuilt)}:[/yellow]")
+        for f in extra_in_rebuilt[:20]:
+            console.print(f"    [green]+[/green] {f}")
+        if len(extra_in_rebuilt) > 20:
+            console.print(f"    [dim]... and {len(extra_in_rebuilt) - 20} more[/dim]")
+
+    # Compare common files by hash
+    mismatched = []
+    for rel_path in common:
+        if original_files[rel_path] != rebuilt_files[rel_path]:
+            mismatched.append(rel_path)
+
+    # Filter mismatched: ignore package.json fields that change between installs
+    real_mismatches = []
+    for rel_path in mismatched:
+        if rel_path.name == "package.json":
+            orig_text = (original_dir / rel_path).read_text(errors="replace")
+            rebuilt_text = (rebuilt_dir / rel_path).read_text(errors="replace")
+            install_fields = {"_resolved", "_integrity", "_from", "_where", "_id",
+                              "_requested", "_requiredBy", "_shasum", "_spec",
+                              "_phantomChildren", "_inBundle"}
+            try:
+                orig_json = json.loads(orig_text)
+                rebuilt_json = json.loads(rebuilt_text)
+                for field in install_fields:
+                    orig_json.pop(field, None)
+                    rebuilt_json.pop(field, None)
+                if orig_json == rebuilt_json:
+                    continue
+            except (json.JSONDecodeError, ValueError):
+                pass
+        real_mismatches.append(rel_path)
+
+    matched_count = len(common) - len(real_mismatches)
+    if real_mismatches:
+        all_match = False
+        console.print(
+            f"\n  [red]Files with different content — {len(real_mismatches)} of {len(common)}:[/red]"
+        )
+        shown = 0
+        for rel_path in real_mismatches:
+            file_link = link(f"{blob_url}/{rel_path}", str(rel_path))
+            console.print(f"    [red]✗[/red] {file_link}")
+            if shown < 5 and rel_path.suffix == ".js":
+                orig_content = (original_dir / rel_path).read_text(errors="replace")
+                rebuilt_content = (rebuilt_dir / rel_path).read_text(errors="replace")
+                show_colored_diff(rel_path, orig_content, rebuilt_content)
+                shown += 1
+        if len(real_mismatches) > 20:
+            console.print(f"    [dim]... showing first 20 of {len(real_mismatches)}[/dim]")
+    else:
+        console.print(f"\n  [green]✓[/green] All {matched_count} common files match")
+
+    return all_match
