+++ utils/verify_action_build/diff_source.py
@@ -1,0 +1,205 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+"""Source file diffing between approved and new versions."""
+
+import shutil
+from pathlib import Path
+
+from rich.panel import Panel
+
+from .console import console, run, ask_confirm, UserQuit
+from .diff_display import show_colored_diff
+
+
+def diff_approved_vs_new(
+    org: str, repo: str, approved_hash: str, new_hash: str, work_dir: Path,
+    ci_mode: bool = False,
+) -> None:
+    """Diff source files between an approved version and the new version."""
+    console.print()
+    console.rule("[bold]Diff: Approved vs New (source changes)[/bold]")
+
+    approved_dir = work_dir / "approved-src"
+    new_dir = work_dir / "new-src"
+    approved_dir.mkdir(exist_ok=True)
+    new_dir.mkdir(exist_ok=True)
+
+    repo_url = f"https://github.com/{org}/{repo}.git"
+
+    excluded_dirs = {"dist", "node_modules", ".git", ".github", "__tests__", "__mocks__"}
+    lock_files = {
+        "package-lock.json", "yarn.lock", "pnpm-lock.yaml", "bun.lockb",
+        "shrinkwrap.json", "npm-shrinkwrap.json",
+    }
+    source_extensions = {".js", ".ts", ".mjs", ".cjs", ".mts", ".cts", ".json", ".yml", ".yaml"}
+
+    with console.status("[bold blue]Fetching source from both versions...[/bold blue]"):
+        clone_dir = work_dir / "repo-clone"
+        run(
+            ["git", "clone", "--no-checkout", repo_url, str(clone_dir)],
+            capture_output=True,
+        )
+
+        skipped_dirs: set[str] = set()
+
+        for label, commit, out_dir in [
+            ("approved", approved_hash, approved_dir),
+            ("new", new_hash, new_dir),
+        ]:
+            run(
+                ["git", "checkout", commit],
+                capture_output=True,
+                cwd=clone_dir,
+            )
+            for f in clone_dir.rglob("*"):
+                if not f.is_file():
+                    continue
+                rel = f.relative_to(clone_dir)
+                matched = [part for part in rel.parts if part in excluded_dirs]
+                if matched:
+                    skipped_dirs.update(matched)
+                    continue
+                if rel.suffix in source_extensions:
+                    dest = out_dir / rel
+                    dest.parent.mkdir(parents=True, exist_ok=True)
+                    shutil.copy2(f, dest)
+
+    console.print("  [green]✓[/green] Fetched source from both versions")
+
+    test_dirs = {"__tests__", "__mocks__"}
+    skipped_test_dirs = sorted(skipped_dirs & test_dirs)
+
+    approved_files = set()
+    new_files = set()
+    for f in approved_dir.rglob("*"):
+        if f.is_file():
+            approved_files.add(f.relative_to(approved_dir))
+    for f in new_dir.rglob("*"):
+        if f.is_file():
+            new_files.add(f.relative_to(new_dir))
+
+    all_files = sorted(approved_files | new_files)
+
+    if not all_files:
+        console.print("  [yellow]No source files found[/yellow]")
+        return
+
+    skipped_locks = sorted(f for f in all_files if f.name in lock_files)
+    has_skips = bool(skipped_locks) or bool(skipped_test_dirs)
+
+    if has_skips:
+        console.print()
+        console.print("  [bold]The following files/directories are excluded from comparison:[/bold]")
+        if skipped_test_dirs:
+            console.print(f"    [dim]⊘ {', '.join(skipped_test_dirs)} (test files only, not part of the action runtime)[/dim]")
+        if skipped_locks:
+            console.print(f"    [dim]⊘ {len(skipped_locks)} lock file(s) (generated by package managers):[/dim]")
+            for f in skipped_locks:
+                console.print(f"      [dim]- {f}[/dim]")
+        for d in sorted(excluded_dirs - {"dist", "node_modules", ".git"} - test_dirs):
+            if any(d in str(f) for f in all_files):
+                console.print(f"    [dim]⊘ {d}/ (not part of the action runtime)[/dim]")
+
+        if not ci_mode:
+            try:
+                if not ask_confirm("  Proceed with these exclusions?"):
+                    console.print("  [yellow]Aborted by user[/yellow]")
+                    return
+            except UserQuit:
+                console.print("  [yellow]Aborted by user[/yellow]")
+                return
+
+    skipped_by_user: list[tuple[Path, str]] = []
+    quit_all = False
+
+    for rel_path in all_files:
+        if rel_path.name in lock_files:
+            continue
+
+        if quit_all:
+            skipped_by_user.append((rel_path, "skipped (quit)"))
+            continue
+
+        approved_file = approved_dir / rel_path
+        new_file = new_dir / rel_path
+
+        if rel_path not in approved_files:
+            console.print(f"  [cyan]+[/cyan] {rel_path} [dim](new file)[/dim]")
+            new_content = new_file.read_text(errors="replace")
+            result = show_colored_diff(rel_path, "", new_content, from_label="approved", to_label="new", border="cyan", ci_mode=ci_mode)
+            if result == "skip_file":
+                skipped_by_user.append((rel_path, "new file"))
+            elif result == "quit":
+                quit_all = True
+            continue
+
+        if rel_path not in new_files:
+            console.print(f"  [cyan]-[/cyan] {rel_path} [dim](removed)[/dim]")
+            approved_content = approved_file.read_text(errors="replace")
+            result = show_colored_diff(rel_path, approved_content, "", from_label="approved", to_label="new", border="cyan", ci_mode=ci_mode)
+            if result == "skip_file":
+                skipped_by_user.append((rel_path, "removed"))
+            elif result == "quit":
+                quit_all = True
+            continue
+
+        approved_content = approved_file.read_text(errors="replace")
+        new_content = new_file.read_text(errors="replace")
+
+        if approved_content == new_content:
+            console.print(f"  [green]✓[/green] {rel_path} [green](identical)[/green]")
+        else:
+            console.print(f"  [cyan]~[/cyan] {rel_path} [cyan](changed — expected between versions)[/cyan]")
+            result = show_colored_diff(rel_path, approved_content, new_content, from_label="approved", to_label="new", border="cyan", ci_mode=ci_mode)
+            if result == "skip_file":
+                skipped_by_user.append((rel_path, "changed"))
+            elif result == "quit":
+                quit_all = True
+
+    console.print()
+
+    if has_skips:
+        excluded_summary = []
+        if skipped_test_dirs:
+            excluded_summary.append(f"  {', '.join(skipped_test_dirs)}/ (test files)")
+        if skipped_locks:
+            for f in skipped_locks:
+                excluded_summary.append(f"  {f} (lock file)")
+        for d in sorted(excluded_dirs - {"dist", "node_modules", ".git"} - test_dirs):
+            if any(d in str(f) for f in all_files):
+                excluded_summary.append(f"  {d}/ (not part of action runtime)")
+        if excluded_summary:
+            console.print(
+                Panel(
+                    "\n".join(excluded_summary),
+                    title="[green bold]Excluded from comparison (confirmed by reviewer)[/green bold]",
+                    border_style="green",
+                    padding=(0, 1),
+                )
+            )
+
+    if skipped_by_user:
+        console.print(
+            Panel(
+                "\n".join(f"  - {f} ({reason})" for f, reason in skipped_by_user),
+                title="[yellow bold]Files skipped — still need manual review[/yellow bold]",
+                border_style="yellow",
+                padding=(0, 1),
+            )
+        )
